Week 6 Task 2 Completed

+++ Week 06/Backend/controller/GoalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../model/goalModel');
+const User = require('../model/userModel');
 
 // Get Method
 const GetGoal = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json(goals);
 })
 
@@ -16,7 +17,8 @@ const SetGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal);
@@ -29,6 +31,19 @@ const UpdateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not founded!');
+    }
+
+    const user = await User.findById(req.user.id);
+    // Check the user
+    if (!user) {
+        res.status(401);
+        throw new Error('User Not Founded!');
+    }
+
+    // Make sure the same user goal updates
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorized!');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -44,6 +59,19 @@ const DeleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not founded!');
+    }
+
+    const user = await User.findById(req.user.id);
+    // Check the user
+    if (!user) {
+        res.status(401);
+        throw new Error('User Not Founded!');
+    }
+
+    // Make sure the same user goal updates
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorized!');
     }
 
     await Goal.findByIdAndDelete(req.params.id);
